Add a BaseValidationError abstract class and move error-related stuff to a new errors module
