+++ valideer/__init__.py
@@ -1,2 +1,3 @@
 from .base import *
-from .validators import  *
+from .errors import *
+from .validators import *
